Exception Modified from Double BigDecimal

+++ src/main/java/com/cupflow/CupFlow_ERP/common/exception/InsufficientStockException.java
@@ -2,9 +2,11 @@ package com.cupflow.CupFlow_ERP.common.exception;
 
 import org.springframework.http.HttpStatus;
 
+import java.math.BigDecimal;
+
 public class InsufficientStockException extends AppException {
 
-    public InsufficientStockException(String materialType, double required, double available) {
+    public InsufficientStockException(String materialType, BigDecimal required, BigDecimal available) {
         super(HttpStatus.CONFLICT,
                 "Insufficient stock for Material " + materialType +
                         " , Required: " + required +
